+++ term/src/terminalstate/kitty.rs
@@ -885,11 +885,16 @@ impl TerminalState {
                 }
             }
 
-            let mut b64_encoded = String::new();
-            for data in data.into_iter() {
-                match data {
+            let mut b64_decoded = vec![];
+            for mut data in data.into_iter() {
+                match &mut data {
+                    KittyImageData::DirectBin(b) => {
+                        b64_decoded.append(b);
+                    }
                     KittyImageData::Direct(b) => {
-                        b64_encoded.push_str(&b);
+                        if !b.is_empty() {
+                            b64_decoded.append(&mut data.load_data()?);
+                        }
                     }
                     data => {
                         anyhow::bail!("expected data chunks to be Direct data, found {:#?}", data)
@@ -897,7 +902,7 @@ impl TerminalState {
                 }
             }
 
-            trans.data = KittyImageData::Direct(b64_encoded);
+            trans.data = KittyImageData::DirectBin(b64_decoded);
 
             if let Some(placement) = place {
                 Ok(KittyImage::TransmitDataAndDisplay {